logger added

+++ network.cpp
@@ -14,7 +14,8 @@ network network::from_string(const std::string& net_str)
     if(cm.size() != 6)
         throw std::invalid_argument("invalid network address");
 
-    uint32_t net_addr = std::stoi(cm[1]) << 24 | std::stoi(cm[2]) << 16 | std::stoi(cm[3]) << 8 | std::stoi(cm[4]);
+    uint32_t net_addr = static_cast<uint32_t>(std::stoi(cm[1]) << 24 | std::stoi(cm[2]) << 16 |
+                                              std::stoi(cm[3]) << 8 | std::stoi(cm[4]));
     uint32_t net_mask = 0;
     int i;
     for(i = 0; i < std::stoi(cm[5]); ++i)
